Move binrelClass to functions/mclass.js
