+++ src/functions/mclass.js
@@ -3,6 +3,7 @@ import defineFunction, {ordargument} from "../defineFunction";
 import buildCommon from "../buildCommon";
 import mathMLTree from "../mathMLTree";
 import ParseNode from "../ParseNode";
+import type {AnyParseNode} from "../ParseNode";
 
 import * as html from "../buildHTML";
 import * as mml from "../buildMathML";
@@ -41,6 +42,20 @@ defineFunction({
     mathmlBuilder,
 });
 
+export const binrelClass = (arg: AnyParseNode) => {
+    // \binrel@ spacing varies with (bin|rel|ord) of the atom in the argument.
+    // (by rendering separately and with {}s before and after, and measuring
+    // the change in spacing).  We'll do roughly the same by detecting the
+    // atom type directly.
+    const atomType = (arg.type === "ordgroup" &&
+        arg.value.length ? arg.value[0].type : arg.type);
+    if (/^(bin|rel)$/.test(atomType)) {
+        return "m" + atomType;
+    } else {
+        return "mord";
+    }
+};
+
 // \@binrel{x}{y} renders like y but as mbin/mrel/mord if x is mbin/mrel/mord.
 // This is equivalent to \binrel@{x}\binrel@@{y} in AMSTeX.
 defineFunction({
@@ -52,7 +67,7 @@ defineFunction({
     handler({parser}, [baseArg, renderArg]) {
         return new ParseNode("mclass", {
             type: "mclass",
-            mclass: buildCommon.binrelClass(baseArg),
+            mclass: binrelClass(baseArg),
             value: [renderArg],
         }, parser.mode);
     },
@@ -72,7 +87,7 @@ defineFunction({
         let mclass;
         if (funcName !== "\\stackrel") {
             // LaTeX applies \binrel spacing to \overset and \underset.
-            mclass = buildCommon.binrelClass(baseArg);
+            mclass = binrelClass(baseArg);
         } else {
             mclass = "mrel";  // for \stackrel
         }
